+++ packages/firestore/test/integration/api/database.test.ts
@@ -23,7 +23,8 @@ import {
   apiDescribe,
   withTestCollection,
   withTestDb,
-  withTestDoc
+  withTestDoc,
+  withTestDocAndInitialData
 } from '../util/helpers';
 import { query } from '../../util/api_helpers';
 
@@ -204,6 +205,108 @@ apiDescribe('Database', persistence => {
           expect(docSnapshot.exists).to.be.ok;
           expect(docSnapshot.data()).to.deep.equal(finalData);
         });
+    });
+  });
+
+  it("can't specify a field mask for a missing field using set", () => {
+    return withTestDoc(persistence, async docRef => {
+      expect(() => {
+        docRef.set(
+          { desc: 'NewDescription' },
+          { mergeFields: ['desc', 'owner'] }
+        );
+      }).to.throw(
+        "Field 'owner' is specified in your field mask but missing from your input data."
+      );
+    });
+  });
+
+  it('can set a subset of fields using a field mask', () => {
+    const initialData = {
+      desc: 'Description',
+      owner: { name: 'Jonny', email: 'abc@xyz.com' }
+    };
+    const finalData = { desc: 'Description', owner: 'Sebastian' };
+    return withTestDocAndInitialData(persistence, initialData, async docRef => {
+      await docRef.set(
+        { desc: 'NewDescription', owner: 'Sebastian' },
+        { mergeFields: ['owner'] }
+      );
+      const result = await docRef.get();
+      expect(result.data()).to.deep.equal(finalData);
+    });
+  });
+
+  it("doesn't apply field delete outside of mask", () => {
+    const initialData = {
+      desc: 'Description',
+      owner: { name: 'Jonny', email: 'abc@xyz.com' }
+    };
+    const finalData = { desc: 'Description', owner: 'Sebastian' };
+    return withTestDocAndInitialData(persistence, initialData, async docRef => {
+      await docRef.set(
+        { desc: firebase.firestore.FieldValue.delete(), owner: 'Sebastian' },
+        { mergeFields: ['owner'] }
+      );
+      const result = await docRef.get();
+      expect(result.data()).to.deep.equal(finalData);
+    });
+  });
+
+  it("doesn't apply field transform outside of mask", () => {
+    const initialData = {
+      desc: 'Description',
+      owner: { name: 'Jonny', email: 'abc@xyz.com' }
+    };
+    const finalData = { desc: 'Description', owner: 'Sebastian' };
+    return withTestDocAndInitialData(persistence, initialData, async docRef => {
+      await docRef.set(
+        {
+          desc: firebase.firestore.FieldValue.serverTimestamp(),
+          owner: 'Sebastian'
+        },
+        { mergeFields: ['owner'] }
+      );
+      const result = await docRef.get();
+      expect(result.data()).to.deep.equal(finalData);
+    });
+  });
+
+  it('can set an empty field mask', () => {
+    const initialData = {
+      desc: 'Description',
+      owner: { name: 'Jonny', email: 'abc@xyz.com' }
+    };
+    const finalData = initialData;
+    return withTestDocAndInitialData(persistence, initialData, async docRef => {
+      await docRef.set(
+        { desc: 'NewDescription', owner: 'Sebastian' },
+        { mergeFields: [] }
+      );
+      const result = await docRef.get();
+      expect(result.data()).to.deep.equal(finalData);
+    });
+  });
+
+  it('can specify fields multiple times in a field mask', () => {
+    const initialData = {
+      desc: 'Description',
+      owner: { name: 'Jonny', email: 'abc@xyz.com' }
+    };
+    const finalData = {
+      desc: 'Description',
+      owner: { name: 'Sebastian', email: 'new@xyz.com' }
+    };
+    return withTestDocAndInitialData(persistence, initialData, async docRef => {
+      await docRef.set(
+        {
+          desc: 'NewDescription',
+          owner: { name: 'Sebastian', email: 'new@xyz.com' }
+        },
+        { mergeFields: ['owner.name', 'owner', 'owner'] }
+      );
+      const result = await docRef.get();
+      expect(result.data()).to.deep.equal(finalData);
     });
   });