+++ packages/firestore/CHANGELOG.md
@@ -21,6 +21,8 @@
   Query.get() should fetch from server only, (by passing { source: 'server' }),
   cache only (by passing { source: 'cache' }), or attempt server and fall back
   to the cache (which was the only option previously, and is now the default).
+- [feature] Added new `{ mergeFields: (string|FieldPath)[] }` option to `set()` 
+  which allows merging of a reduced subset of fields.
 
 # 0.3.6
 - [fixed] Fixed a regression in the Firebase JS release 4.11.0 that could